Make Steal and the Less File Manager share a common source for fetching
Since Steal and Less have their own separate fetching mechanism it means
they also have their own separate caching mechanisms. This change adds a
`.fetch` hook for Steal and a `.doXHR` hook for Less that share a common
cache, so if a resource has already been fetched in one it will not be
refetched in the other. Closes #18